automatically backed up by learn

+++ index.js
@@ -28,5 +28,5 @@ function makeInt(n){
 }
 //ninth
 function  preserveDecimal(n){
-  return parseInt(n,10)
+  return parseFloat(n,10)
 }
